spolit categories and types

+++ contracts/nft/src/user_info.rs
@@ -1,7 +1,7 @@
 use soroban_sdk::{Address, Env};
 use crate::storage_types::DataKey;
 
-pub fn read_user_level(e: &Env, user: Address) -> u64 {
+pub fn read_user_level(e: &Env, user: Address) -> u32 {
     e.storage()
         .persistent()
         .get(&DataKey::UserLevel(user))
